requirements.txt
django
django-rest-framework
django-cors-headers
djoser
pillow
stripe

